requirements.txt
streamlit
python-docx
langchain
langchain-google-genai
langchain-community
pdfminer.six
chromadb
google-generativeai